+++ types/app-state.d.ts
@@ -2,8 +2,8 @@ import { Persist } from './persist'
 
 export type AppStateStoreAction = (payload?: any) => void
 
-export interface AppStateStoreOfStores<TS> {
-  [key: string]: AppStateStoreClass<TS>
+export interface AppStateStoreOfStores<TS, TP> {
+  [key: string]: AppStateStoreClass<TS, TP>
 }
 
 export interface AppStateStoreBaseOptions<TS> {
@@ -16,16 +16,18 @@ export type AppStateStoreOptions<TS> = AppStateStoreBaseOptions<TS> & {
   [key: string]: AppStateStoreAction
 }
 
-export type AppStateStore<TS> = AppStateStoreOfStores<TS> | AppStateStoreClass<TS>
+export type AppStateStore<TS, TP> = AppStateStoreOfStores<TS, TP> | AppStateStoreClass<TS, TP>
 
-declare class AppStateStoreClass<TS> {
+export type AppStateGetStateFunc<TS, TP> = (state: TS) => TP
+
+declare class AppStateStoreClass<TS, TP> {
   constructor(options: AppStateStoreOptions<TS>)
   setState (setter: (state: TS) => any, callback?: () => void): void
-  getState (key: string): any
+  getState (key: string | AppStateGetStateFunc<TS, TP>): TP
   subscribe (callback: () => any): void
   clear (notifySubscribers?: boolean): void
 }
 
-export declare function createAppState<TS> (key: string, options: AppStateStoreOptions<TS>): AppStateStoreClass<TS>
+export declare function createAppState<TS, TP> (key: string, options: AppStateStoreOptions<TS>): AppStateStoreClass<TS, TP>
 
-export declare function getAppState<TS> (key: string): AppStateStoreClass<TS>
+export declare function getAppState<TS, TP> (key: string): AppStateStoreClass<TS, TP>