+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2022-02-22
+
+### Updates
+- Allow functions to be passed to `getState`
+- Memoize `getState` function return values
+
 ## [1.5.0] - 2022-02-16
 
 ### Updates